+++ packages/cinterop/src/jvm/kotlin/io/realm/kotlin/internal/interop/RealmInterop.kt
@@ -19,6 +19,7 @@ package io.realm.kotlin.internal.interop
 import io.realm.kotlin.internal.interop.Constants.ENCRYPTION_KEY_LENGTH
 import io.realm.kotlin.internal.interop.sync.AuthProvider
 import io.realm.kotlin.internal.interop.sync.CoreSubscriptionSetState
+import io.realm.kotlin.internal.interop.sync.CoreSyncSessionState
 import io.realm.kotlin.internal.interop.sync.CoreUserState
 import io.realm.kotlin.internal.interop.sync.JVMSyncSessionTransferCompletionCallback
 import io.realm.kotlin.internal.interop.sync.MetadataMode
@@ -859,6 +860,18 @@ actual object RealmInterop {
             syncSession.cptr(),
             JVMSyncSessionTransferCompletionCallback(callback)
         )
+    }
+
+    actual fun realm_sync_session_state(syncSession: RealmSyncSessionPointer): CoreSyncSessionState {
+        return CoreSyncSessionState.of(realmc.realm_sync_session_get_state(syncSession.cptr()))
+    }
+
+    actual fun realm_sync_session_pause(syncSession: RealmSyncSessionPointer) {
+        realmc.realm_sync_session_pause(syncSession.cptr())
+    }
+
+    actual fun realm_sync_session_resume(syncSession: RealmSyncSessionPointer) {
+        realmc.realm_sync_session_resume(syncSession.cptr())
     }
 
     @Suppress("LongParameterList")