+++ packages/library-sync/src/commonMain/kotlin/io/realm/kotlin/mongodb/sync/SyncSession.kt
@@ -69,6 +69,32 @@ public interface SyncSession {
     public suspend fun uploadAllLocalChanges(timeout: Duration = Duration.INFINITE): Boolean
 
     /**
+     * The current session state. See [State] for more details about each state.
+     */
+    public val state: State
+
+    /**
+     * Pauses synchronization with Atlas until the Realm is closed and re-opened again.
+     *
+     * Synchronization can also be re-activated by calling [resume].
+     *
+     * If the session is already [State.INACTIVE], calling this method will do nothing.
+     */
+    public fun pause()
+
+    /**
+     * Attempts to resume the session and activate synchronization with Atlas.
+     *
+     * This happens automatically when opening the Realm, so doing it manually should only
+     * be needed if the session was paused using [pause].
+     *
+     * If the session was already [State.ACTIVE], calling this method will do nothing.
+     *
+     * If the session state is [State.DYING], the session will be moved back to [State.ACTIVE].
+     */
+    public fun resume()
+
+    /**
      * Interface used to report any session errors.
      *
      * @see SyncConfiguration.Builder.errorHandler
@@ -84,5 +110,45 @@ public interface SyncSession {
          * @param error the [SyncException] being reported by the server.
          */
         public fun onError(session: SyncSession, error: SyncException)
+    }
+
+    /**
+     * The possible states for [SyncSession] to be.
+     *
+     * A Realm will automatically synchronize data with the server if the session is either [ACTIVE]
+     * or [DYING] and is connected.
+     */
+    public enum class State {
+        /**
+         * This is the initial state. The session is closed. No data is being synchronized. The session
+         * will automatically transition to [ACTIVE] when a Realm is opened.
+         */
+        INACTIVE,
+
+        /**
+         * The Realm is open and data will be synchronized between the device and the server
+         * if the underlying connection is connected.
+         *
+         * The session will remain in this state until the Realm
+         * is closed. In which case it will become [DYING].
+         */
+        ACTIVE,
+
+        /**
+         * The Realm was closed, but still contains data that needs to be synchronized to the server.
+         * The session will attempt to upload all local data before going [INACTIVE].
+         */
+        DYING,
+
+        /**
+         * The user is attempting to synchronize data but needs a valid access token to do so. Realm
+         * will either use a cached token or automatically try to acquire one based on the current
+         * users login. This requires a network connection.
+         *
+         * Data cannot be synchronized in this state.
+         *
+         * Once a valid token is acquired, the session will transition to [ACTIVE].
+         */
+        WAITING_FOR_ACCESS_TOKEN
     }
 }